requirements.txt
streamlit
pdfplumber
openai
sqlalchemy
pandas
python-dotenv
PyMuPDF
